requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
billiard==3.6.4.0
bleach==5.0.1
celery==5.2.7
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
Deprecated==1.2.13
dj-rest-auth==2.2.5
Django==3.2
django-debug-toolbar==3.4.0
django-filter==22.1
django-redis==5.2.0
django-rest-swagger==2.2.0
django-seed==0.3.1
django-summernote==0.8.20.0
django-taggit==3.0.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
Faker==13.15.1
idna==3.3
importlib-metadata==4.12.0
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
Markdown==3.4.1
MarkupSafe==2.1.1
openapi-codec==1.3.2
packaging==21.3
Pillow==9.2.0
prompt-toolkit==3.0.31
psycopg2-binary==2.9.3
PyJWT==2.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
redis==4.3.4
requests==2.28.1
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
toposort==1.7
uritemplate==4.1.1
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
wrapt==1.14.1
zipp==3.8.1
